Add guards to `eachSystem' to prevent NixOS attrs being restructured
Unlike `defaultPackage' and `legacyPackages' the `nixosModule(s)' and
`nixosConfiguration' attributes are not structued by system.

e.g.
nixosModules."module"

This patch adds some checks on the name of the attribute in `eachSystem'
and avoids restructuring the value of it if the name matches one of the
ones previously mentioned, otherwise its behaviour is unchanged.